reintegrate procman branch
Allow sheriff to control the stop signal that deputies send to commands to
request a clean exit, as well as the time a command is allowed to cleanly exit
before the deputy sends a SIGKILL.  The stop signal defaults to SIGINT, but can
also be set to SIGTERM or SIGKILL for now.

Add new LCM message types to support new controls.  New sheriff and deputy
implementations should be backwards compatible with old sheriffs and deputies.